pyproject.toml
[project]
name = "ysth-junior-ai-engineer-test-mar2026"
version = "0.1.0"
description = "Thai Securities Q&A - Intelligent financial research assistant"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi==0.135.3",
    "uvicorn[standard]==0.44.0",
    "python-dotenv==1.2.2",
    "pydantic==2.12.5",
    "faiss-cpu==1.13.2",
    "numpy==2.4.4",
    "scikit-learn==1.8.0",
    "scipy==1.17.1",
    "openai==2.30.0",
    "streamlit==1.56.0",
    "requests==2.33.1",
    "google-genai>=1.70.0",
    "rank-bm25>=0.2.2",
]
